requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
ollama
telethon
transformers
torch
google-generativeai>=0.7
